+++ src/resources/views/layouts/default/private.blade.php
@@ -8,7 +8,7 @@
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>@yield('title') &middot; {{ setting('appshell.ui.name') }}</title>
+    <title>@yield('title') &middot; {{ $appshell->name }}</title>
 
     <!-- Styles -->
     <link href="https://fonts.googleapis.com/css?family=Montserrat:300,300i,400,400i,500,500i,600,600i,700,700i" media="all" type="text/css" rel="stylesheet" />
@@ -42,7 +42,7 @@
 
     <footer class="app-footer">
         @section('footer')
-        &copy; {{ date('Y') }}&nbsp;<a href="{{ $appshell->url }}">{{ setting('appshell.ui.name') }}</a>
+        &copy; {{ date('Y') }}&nbsp;<a href="{{ $appshell->url }}">{{ $appshell->name }}</a>
         @endsection
         @yield('footer')
     </footer>