+++ src/resources/manifest.php
@@ -2,5 +2,5 @@
 
 return [
     'name'    => 'Konekt AppShell Box',
-    'version' => '2.0-dev'
+    'version' => '2.0.0'
 ];